+++ src/main/scala/edu/ie3/simona/service/em/EmServiceCore.scala
@@ -6,65 +6,114 @@
 
 package edu.ie3.simona.service.em
 
-import edu.ie3.datamodel.models.value.{PValue, SValue}
 import edu.ie3.simona.agent.em.EmAgent
 import edu.ie3.simona.api.FlexConversion
-import edu.ie3.simona.api.data.model.em.{FlexOptions, SetPoint}
+import edu.ie3.simona.api.FlexConversion.{convert, convertOptions}
+import edu.ie3.simona.api.data.connection.ExtEmDataConnection
+import edu.ie3.simona.api.data.connection.ExtEmDataConnection.EmMode
+import edu.ie3.simona.api.data.model.em.{
+  DisaggregatedFlexOptions,
+  EmCommunicationMessage,
+  EmData,
+  FlexOptionRequest,
+  SetPoint,
+  FlexOptions as ExtFlexOptions,
+}
 import edu.ie3.simona.api.ontology.em.*
+import edu.ie3.simona.exceptions.CriticalFailureException
 import edu.ie3.simona.ontology.messages.ServiceMessage.{
   EmFlexMessage,
   EmServiceRegistration,
   ServiceResponseMessage,
 }
+import edu.ie3.simona.ontology.messages.flex.FlexType.PowerLimit
 import edu.ie3.simona.ontology.messages.flex.FlexibilityMessage.*
-import edu.ie3.simona.util.ReceiveDataMap
-import edu.ie3.simona.util.SimonaConstants.INIT_SIM_TICK
-import edu.ie3.util.quantities.QuantityUtils.asMegaWatt
-import edu.ie3.util.scala.quantities.QuantityConversionUtils.toSquants
+import edu.ie3.simona.service.DataTimeType.Current
+import edu.ie3.simona.service.em.EmServiceCore.EmAgentState
+import edu.ie3.simona.util.CollectionUtils.asJava
+import edu.ie3.simona.util.SimonaConstants.{
+  FIRST_TICK_IN_SIMULATION,
+  INIT_SIM_TICK,
+}
+import edu.ie3.simona.util.{ReceiveDataMap, ReceiveMultiDataMap}
 import org.apache.pekko.actor.typed.ActorRef
 import org.slf4j.Logger
-import squants.Power
-import tech.units.indriya.ComparableQuantity
 
 import java.time.ZonedDateTime
-import java.util.{Optional, OptionalLong, UUID}
-import javax.measure.quantity.Power as PsdmPower
-import scala.jdk.CollectionConverters.MapHasAsJava
-import scala.jdk.OptionConverters.{RichOption, RichOptional, RichOptionalLong}
+import java.util.{OptionalLong, UUID}
+import scala.collection.mutable
+import scala.jdk.CollectionConverters.{
+  CollectionHasAsScala,
+  MapHasAsJava,
+  MapHasAsScala,
+}
+import scala.jdk.OptionConverters.RichOptionalLong
+import scala.math.max
+import scala.util.Try
 
-/** Trait for all em service cores.
+/** Basic service core for an [[ExtEmDataService]].
+  *
+  * @param uuidToAgent
+  *   Map: uuid to em agent reference.
+  * @param agentToUuid
+  *   Map: em agent reference to uuid.
+  * @param uncontrolled
+  *   A set of uuids of uncontrolled em models.
+  * @param uuidToInferior
+  *   A map that contains information about uuids of inferior em agents. This
+  *   information is used to determine the disaggregated flex options.
+  * @param uuidToParent
+  *   A map: uuid to parent uuid.
+  * @param completions
+  *   ReceiveDataMap: uuid to completions.
+  * @param nextActivation
+  *   A map: uuid to next activation tick.
+  * @param allFlexOptions
+  *   A map: uuid to flex options
+  * @param emStates
+  *   A map: uuid to em agent state.
+  * @param emDataStore
+  *   ReceiveMultiDataMap: uuid to flex option.
+  * @param internal
+  *   A set of uuids of models that simulated internally.
+  * @param sendDataToExt
+  *   True, if em data should be sent to the external simulation.
+  * @param canHandleSetPoints
+  *   True, if the core can sent the received em set points to the agent. It
+  *   will only be true, of all em agent are activated for the current tick and
+  *   therefore able to process the send set points.
+  * @param setPointOption
+  *   Option for em set points that needs to be handled at a later time.
   */
-trait EmServiceCore {
-
-  /** Map: uuid to em agent reference.
-    */
-  val uuidToAgent: Map[UUID, ActorRef[EmAgent.Message]]
-
-  val agentToUuid: Map[ActorRef[FlexRequest] | ActorRef[FlexResponse], UUID]
-
-  val uncontrolled: Set[UUID]
-
-  val uuidToInferior: Map[UUID, Set[UUID]]
-
-  val uuidToParent: Map[UUID, UUID]
-
-  /** ReceiveDataMap: uuid to completions.
-    */
-  val completions: ReceiveDataMap[UUID, FlexCompletion]
-
-  val nextActivation: Map[UUID, Long]
-
-  /** Extension to convert a squants power value to a psdm power value.
-    */
-  extension (value: Power) {
-    def toQuantity: ComparableQuantity[PsdmPower] = value.toMegawatts.asMegaWatt
-  }
+case class EmServiceCore(
+    mode: ExtEmDataConnection.EmMode,
+    emUnitsToRegister: Set[UUID],
+    sendDataToExt: Boolean = false,
+    uuidToAgent: Map[UUID, ActorRef[EmAgent.Message]] = Map.empty,
+    agentToUuid: Map[ActorRef[FlexRequest] | ActorRef[FlexResponse], UUID] =
+      Map.empty,
+    uncontrolled: Set[UUID] = Set.empty,
+    uuidToInferior: Map[UUID, Set[UUID]] = Map.empty,
+    uuidToParent: Map[UUID, UUID] = Map.empty,
+    completions: ReceiveDataMap[UUID, FlexCompletion] = ReceiveDataMap.empty,
+    nextActivation: Map[UUID, Long] = Map.empty,
+    allFlexOptions: Map[UUID, ExtFlexOptions] = Map.empty,
+    emStates: Map[UUID, EmAgentState] = Map.empty,
+    emDataStore: ReceiveMultiDataMap[UUID, EmData] = ReceiveMultiDataMap.empty,
+    internal: Set[UUID] = Set.empty,
+    canHandleSetPoints: Boolean = false,
+    setPointOption: Option[Map[UUID, SetPoint]] = None,
+) {
 
   given Conversion[OptionalLong, Option[Long]] =
     (x: OptionalLong) => x.toScala
   given Conversion[Option[Long], OptionalLong] = {
     case Some(value) => OptionalLong.of(value)
     case None        => OptionalLong.empty
+  }
+
+  final def init(): Unit = {
+    uncontrolled.foreach(uuidToAgent(_) ! FlexInit(PowerLimit, Current))
   }
 
   /** Method to handle a registration message.
@@ -76,7 +125,52 @@ trait EmServiceCore {
     */
   def handleRegistration(
       emServiceRegistration: EmServiceRegistration
-  ): EmServiceCore
+  ): EmServiceCore = {
+    val uuid = emServiceRegistration.inputUuid
+    val ref = emServiceRegistration.requestingActor
+
+    val (
+      updatedUncontrolled,
+      updatedInferior,
+      updatedUuidToParent,
+      updatedCompletions,
+    ) =
+      emServiceRegistration.parentUuid match {
+        case Some(parent) =>
+          val inferior = uuidToInferior.get(parent) match {
+            case Some(inferiorUuids) =>
+              inferiorUuids ++ Seq(uuid)
+            case None =>
+              Set(uuid)
+          }
+
+          (
+            uncontrolled,
+            uuidToInferior.updated(parent, inferior),
+            uuidToParent.updated(uuid, parent),
+            completions,
+          )
+        case None =>
+          (
+            uncontrolled + uuid,
+            uuidToInferior,
+            uuidToParent,
+            completions.addExpectedKey(uuid),
+          )
+      }
+
+    copy(
+      emUnitsToRegister = emUnitsToRegister.excl(uuid),
+      uuidToAgent = uuidToAgent.updated(uuid, ref),
+      agentToUuid = agentToUuid.updated(ref, uuid),
+      uncontrolled = updatedUncontrolled,
+      uuidToInferior = updatedInferior,
+      uuidToParent = updatedUuidToParent,
+      completions = updatedCompletions,
+      nextActivation = nextActivation.updated(uuid, 0),
+      emStates = emStates.updated(uuid, EmAgentState()),
+    )
+  }
 
   /** Method to handle the received message from the external simulation.
     * @param tick
@@ -94,7 +188,115 @@ trait EmServiceCore {
       extMsg: EmDataMessageFromExt,
   )(using
       log: Logger
-  ): (EmServiceCore, Option[EmDataResponseMessageToExt])
+  ): (EmServiceCore, Option[EmDataResponseMessageToExt]) = extMsg match {
+    case internal: EmSimulationInternal =>
+      // the service should simulate the tick internal
+      val internalTick = internal.tick
+
+      val uuids = uncontrolled
+        .filter { uuid => nextActivation(uuid) == internalTick }
+        .map { uuid =>
+          uuidToAgent(uuid) ! FlexActivation(tick)
+          uuid
+        }
+
+      (
+        copy(
+          completions = completions.addExpectedKeys(uuids),
+          emDataStore = emDataStore.addExpectedKeys(uuids),
+          internal = uuids,
+        ),
+        None,
+      )
+
+    case requestEmCompletion: RequestEmCompletion =>
+      // finish tick and return next tick
+      val extTick = requestEmCompletion.tick
+
+      if extTick != tick then {
+        throw new CriticalFailureException(
+          s"Received completion request for tick '$extTick', while being in tick '$tick'."
+        )
+      } else {
+        log.info(s"Request to finish for tick '$tick' received.")
+
+        val nextTick: OptionalLong = if emStates.exists(_._2.isActivated) then {
+          requestEmCompletion.maybeNextTick
+        } else getMaybeNextTick(tick)
+
+        (
+          this,
+          Some(new EmCompletion(nextTick)),
+        )
+      }
+
+    case provideEmData: ProvideEmData =>
+      if !provideEmData.flexOptions.isEmpty then {
+        log.warn(
+          s"We received the following data '$provideEmData'. The base service can currently not handle the provided flex options."
+        )
+      }
+
+      val agents = if tick == 0 then {
+        uuidToAgent
+      } else {
+        uuidToAgent.filter { case (uuid, _) => nextActivation(uuid) <= tick }
+      }
+
+      val flexRequests = provideEmData.flexRequests.asScala.flatMap {
+        case (entity, request) =>
+          uuidToAgent.get(entity).map { ref =>
+            ref ! FlexActivation(tick)
+            entity
+          }
+      }.toSet
+
+      val updatedState = copy(
+        emDataStore = ReceiveMultiDataMap(flexRequests),
+        completions = completions.addExpectedKeys(flexRequests),
+        sendDataToExt = flexRequests.nonEmpty,
+      )
+
+      // handle set points
+      val setPoints = provideEmData.setPoints().asScala.toMap
+
+      if setPoints.nonEmpty then {
+
+        if canHandleSetPoints then {
+          handleSetPoint(tick, setPoints, log)
+
+          (updatedState, None)
+        } else {
+          val entities = setPoints.keySet
+
+          entities.foreach { entity =>
+            uuidToAgent.get(entity) match {
+              case Some(ref) =>
+                // activate the necessary em agent, this is needed, because an em agent needs to know
+                // its current flex option to properly handle the given set point
+                ref ! FlexActivation(tick)
+              case None =>
+                log.warn(s"Received entity: $entity")
+            }
+          }
+
+          (
+            updatedState.copy(
+              emDataStore = updatedState.emDataStore.addExpectedKeys(entities),
+              completions = updatedState.completions.addExpectedKeys(entities),
+              setPointOption = Some(setPoints),
+            ),
+            None,
+          )
+        }
+
+      } else (updatedState, None)
+
+    case other =>
+      throw new CriticalFailureException(
+        s"The EmServiceBaseCore is not able to handle the message: $extMsg"
+      )
+  }
 
   /** Method to handle data response messages from the em agents.
     * @param tick
@@ -193,7 +395,117 @@ trait EmServiceCore {
       tick: Long,
       flexResponse: FlexResponse,
       receiver: Either[UUID, ActorRef[FlexResponse]],
-  )(using log: Logger): (EmServiceCore, Option[EmDataResponseMessageToExt])
+  )(using log: Logger): (EmServiceCore, Option[EmDataResponseMessageToExt]) = {
+
+    val receiverUuid = receiver match {
+      case Right(ref) =>
+        ref ! flexResponse
+        agentToUuid(ref)
+      case Left(uuid) =>
+        uuid
+    }
+
+    flexResponse match {
+      case scheduleFlexActivation @ ScheduleFlexActivation(
+            modelUuid,
+            _,
+            scheduleKey,
+          ) if tick < FIRST_TICK_IN_SIMULATION =>
+        receiver match {
+          case Left(uuid) =>
+            log.debug(s"Unlocking msg: $scheduleFlexActivation")
+            scheduleFlexActivation.scheduleKey.foreach(_.unlock())
+
+            uuidToAgent(uuid) ! FlexInit(PowerLimit, Current)
+
+          case Right(ref) =>
+            log.debug(s"Forwarding the message to: $ref")
+            ref ! scheduleFlexActivation
+        }
+
+        (this, None)
+
+      case provideFlexOptions: ProvideFlexOptions =>
+        val updated = handleFlexOptions(receiverUuid, provideFlexOptions)
+
+        if updated.isComplete then {
+          // we received all flex options
+
+          val (data, updatedStore) = updated.getFinished
+
+          val updatedCore = copy(
+            emDataStore = updatedStore,
+            canHandleSetPoints = true,
+          )
+
+          if internal.nonEmpty then {
+            internal.map(uuidToAgent).foreach(_ ! IssueNoControl(tick))
+
+            (updatedCore, None)
+
+          } else if sendDataToExt then {
+            val dataToSend = data
+
+            // we have received an option request, that will now be answered
+            (updatedCore, Some(new EmResultResponse(dataToSend.asJava)))
+
+          } else {
+
+            setPointOption match {
+              case Some(setPoints) =>
+                // we have received new set points, that are not handled yet => we will handle them now
+                handleSetPoint(tick, setPoints, log)
+
+                (updatedCore.copy(setPointOption = None), None)
+              case None =>
+                // we are now able to handle set points, but we have not yet received any
+                (updatedCore, None)
+            }
+          }
+
+        } else {
+          log.debug(s"Missing flex options for: ${updated.getExpectedKeys}")
+
+          (copy(emDataStore = updated), None)
+        }
+
+      case completion: FlexCompletion =>
+        val (updated, extMsgOption, nextTick, finished) =
+          handleCompletion(tick, completion)
+
+        if finished then {
+          // the next activations
+          val updatedNextActivation =
+            nextActivation ++ updated.receivedData.flatMap { case (uuid, msg) =>
+              msg.requestAtTick.map(uuid -> _)
+            }
+
+          val updatedStateData = copy(
+            completions = ReceiveDataMap.empty,
+            sendDataToExt = false,
+            canHandleSetPoints = false,
+            nextActivation = updatedNextActivation,
+            internal = Set.empty,
+          )
+
+          val msgToExt = if internal.nonEmpty then {
+            Some(new EmCompletion(updatedStateData.getMaybeNextTick(tick)))
+          } else extMsgOption
+
+          log.info(s"Em service completed for tick: $tick")
+
+          (updatedStateData, msgToExt)
+
+        } else {
+          log.debug(s"Missing completion for: ${updated.getExpectedKeys}")
+
+          (copy(completions = updated), extMsgOption)
+        }
+
+      case _ =>
+        (this, None)
+    }
+  }
 
   /** Method to handle flex requests to the em agents.
     * @param flexRequest
@@ -209,7 +521,39 @@ trait EmServiceCore {
   def handleFlexRequest(
       flexRequest: FlexRequest,
       receiver: ActorRef[FlexRequest],
-  )(using log: Logger): (EmServiceCore, Option[EmDataResponseMessageToExt])
+  )(using log: Logger): (EmServiceCore, Option[EmDataResponseMessageToExt]) = {
+    log.debug(s"$receiver: $flexRequest")
+    receiver ! flexRequest
+
+    val uuid = agentToUuid(receiver)
+    (
+      copy(completions = completions.addExpectedKey(uuid)),
+      None,
+    )
+  }
+
+  /** Method to handle flex options.
+    *
+    * @param receiver
+    *   The receiver of the flex options.
+    * @param provideFlexOptions
+    *   The provided flex options.
+    * @return
+    *   An updated service core and a map: uuid to flex options
+    */
+  private def handleFlexOptions(
+      receiver: UUID,
+      provideFlexOptions: ProvideFlexOptions,
+  ): ReceiveMultiDataMap[UUID, EmData] = provideFlexOptions match {
+    case ProvideFlexOptions(modelUuid: UUID, fo) =>
+      val result = fo.toExt(receiver, modelUuid)
+
+      if emDataStore.expects(modelUuid) then {
+        emDataStore.addData(modelUuid, result)
+      } else emDataStore
+
+    case _ => emDataStore
+  }
 
   /** Method to handle received completion messages.
     * @param tick
@@ -251,12 +595,99 @@ trait EmServiceCore {
     * @return
     *   An option for the next activation tick.
     */
-  final def getMaybeNextTick(tick: Long): Option[Long] = {
+  private final def getMaybeNextTick(tick: Long): Option[Long] = {
     val allActivations = completions.receivedData.flatMap {
       case (_, completion) =>
         completion.requestAtTick
     } ++ nextActivation.values.filter(_ > tick)
 
     allActivations.minOption
+  }
+}
+
+object EmServiceCore {
+
+  final case class EmAgentState(
+      private var receivedActivation: Boolean = false,
+      private var disaggregated: Boolean = false,
+      private val awaitedFlexOptions: mutable.Set[UUID] = mutable.Set.empty,
+      private var awaitedSetPoint: Boolean = false,
+      private var waitingForInternal: Boolean = false,
+      private var waitingForRelease: Boolean = false,
+  ) {
+    def setReceivedRequest(value: Boolean = false): Unit = {
+      receivedActivation = true
+      disaggregated = value
+      waitingForInternal = true
+      awaitedSetPoint = true
+    }
+
+    def setWaitingForRelease(): Unit = {
+      waitingForRelease = true
+    }
+
+    def setReceivedRelease(): Unit = {
+      receivedActivation = true
+      waitingForInternal = false
+      waitingForRelease = false
+    }
+
+    def addSendRequest(request: UUID): Unit = {
+      awaitedFlexOptions.add(request)
+    }
+
+    def handleReceivedFlexOption(flexOption: UUID): Unit = {
+      awaitedFlexOptions.remove(flexOption)
+
+      if awaitedFlexOptions.isEmpty then {
+        waitingForInternal = true
+      } else {
+        waitingForInternal = false
+      }
+    }
+
+    def handleReceivedFlexOptions(flexOptions: Seq[UUID]): Unit = {
+      flexOptions.foreach(awaitedFlexOptions.remove)
+
+      if awaitedFlexOptions.isEmpty then {
+        waitingForInternal = true
+      } else {
+        waitingForInternal = false
+      }
+    }
+
+    def setReceivedSetPoint(): Unit = {
+      awaitedSetPoint = false
+      waitingForInternal = true
+    }
+
+    def setWaitingForInternal(value: Boolean): Unit = {
+      waitingForInternal = value
+    }
+
+    def getAwaited: Set[UUID] = awaitedFlexOptions.toSet
+
+    def sentDisaggregated: Boolean = disaggregated
+
+    def isWaitingForActivation: Boolean = !receivedActivation
+
+    def isWaitingForExtern: Boolean =
+      (awaitedFlexOptions.nonEmpty || awaitedSetPoint) && !waitingForInternal
+
+    def isWaitingForSetPoint: Boolean = awaitedSetPoint
+
+    def isWaitingForRelease: Boolean = waitingForRelease
+
+    def isWaitingForInternal: Boolean = waitingForInternal
+
+    def isActivated: Boolean = receivedActivation
+
+    def clear(): Unit = {
+      receivedActivation = false
+      disaggregated = false
+      awaitedFlexOptions.clear
+      awaitedSetPoint = false
+      waitingForInternal = false
+    }
   }
 }