+++ src/main/scala/edu/ie3/simona/sim/setup/ExtSimSetup.scala
@@ -36,7 +36,7 @@ import scala.util.{Failure, Success, Try}
 
 object ExtSimSetup {
 
-  private val log: Logger = LoggerFactory.getLogger(ExtSimSetup.getClass)
+  private given log: Logger = LoggerFactory.getLogger(ExtSimSetup.getClass)
 
   /** Method to set up all external simulations defined via the given
     * [[ExtLinkInterface]]s.
@@ -88,7 +88,7 @@ object ExtSimSetup {
 
       // creating the data connection
       val extSimDataConnection = new ExtSimDataConnection(extSimAdapter)
-      val setUpData = new SetupData(
+      given setupData: SetupData = new SetupData(
         args,
         config,
         grid,
@@ -98,12 +98,12 @@ object ExtSimSetup {
 
       Try {
         // sets up the external simulation
-        extLink.setup(setUpData)
+        extLink.setup(setupData)
         extLink.getExtSimulation
       }.map { extSimulation =>
         // sets the data connection and the setup data explicitly
         extSimulation.setDataConnection(extSimDataConnection)
-        extSimulation.setSetupData(setUpData)
+        extSimulation.setSetupData(setupData)
 
         // send init data right away, init activation is scheduled
         extSimAdapter ! ExtSimAdapter.Create(
@@ -158,6 +158,7 @@ object ExtSimSetup {
       extSimAdapter: ActorRef[ControlResponseMessageFromExt],
       resultProxy: ActorRef[ResultServiceProxy.Message],
       startTime: ZonedDateTime,
+      apiSetupData: SetupData,
   ): ExtSimSetupData = {
     // the data connections this external simulation provides
     val connections = extSimulation.getDataConnections.asScala
@@ -182,10 +183,18 @@ object ExtSimSetup {
               )
               setupData
             } else {
+              val emUnits =
+                apiSetupData.gridContainer.getEmUnits.getEmUnitsMap.keySet.asScala.toSet
+
               val serviceRef = context.spawn(
                 ExtEmDataService(
                   scheduler,
-                  InitExtEmData(extEmDataConnection, startTime),
+                  InitExtEmData(
+                    scheduler,
+                    extEmDataConnection,
+                    startTime,
+                    emUnits,
+                  ),
                   ScheduleLock.singleKey(context, scheduler, INIT_SIM_TICK),
                 ),
                 "ExtEmDataService",
@@ -196,7 +205,7 @@ object ExtSimSetup {
                 extSimAdapter,
               )
 
-              extSimSetupData.update(extEmDataConnection, serviceRef)
+              setupData.update(extEmDataConnection, serviceRef)
             }
 
           case extEvDataConnection: ExtEvDataConnection =>
@@ -220,7 +229,7 @@ object ExtSimSetup {
               extSimAdapter,
             )
 
-            extSimSetupData.update(extEvDataConnection, serviceRef)
+            setupData.update(extEvDataConnection, serviceRef)
 
           case extResultDataConnection: ExtResultDataConnection =>
             val extResultProvider = context.spawn(
@@ -237,7 +246,7 @@ object ExtSimSetup {
               extSimAdapter,
             )
 
-            extSimSetupData.update(extResultDataConnection, extResultProvider)
+            setupData.update(extResultDataConnection, extResultProvider)
 
           case extResultListener: ExtResultListener =>
             val extResultEventListener = context.spawn(
@@ -248,7 +257,7 @@ object ExtSimSetup {
             // add the external listener to the proxy
             resultProxy ! AddListener(extResultEventListener)
 
-            extSimSetupData.update(extResultListener, extResultEventListener)
+            setupData.update(extResultListener, extResultEventListener)
 
           case otherConnection =>
             log.warn(